feat(ct-metrics): use uint32 for packet counters

+++ pkg/plugin/conntrack/conntrack_bpfel_x86.go
@@ -23,8 +23,8 @@ type conntrackCtEntry struct {
 	ConntrackMetadata  struct {
 		BytesForwardCount   uint64
 		BytesReplyCount     uint64
-		PacketsForwardCount uint64
-		PacketsReplyCount   uint64
+		PacketsForwardCount uint32
+		PacketsReplyCount   uint32
 	}
 }
